requirements.txt
flask~=3.1.2
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
Flask-Mail==0.10.0
python-dotenv==1.1.1
email-validator==2.3.0
gunicorn
psycopg2-binary
xhtml2pdf